Fix : 새로고침해도 유저 데이터 저장되게 수정
