+++ frontend/src/components/Layout/Layout.jsx
@@ -1,8 +1,12 @@
 import { Navbar } from '@components/index';
 import { MY_USER_DATA } from '@src/constants/user';
-
 import styles from './Layout.module.css';
 import classNames from 'classnames/bind';
+import { storeContext } from '@stores/index';
+import { useContext, useReducer, useEffect } from 'react';
+import jwtDecode from 'jwt-decode';
+import { getToken } from '@services/login';
+import { useNavigate } from 'react-router-dom';
 
 const cx = classNames.bind(styles);
 
@@ -14,6 +18,33 @@ export const Layout = ({ children, hideNavbar }) => {
     margin: hideNavbar ? '0' : '0 auto 24px',
     padding: hideNavbar ? '0' : '0 24px',
   };
+
+  const navigate = useNavigate();
+
+  const [user, userDispatch] = useContext(storeContext).user;
+
+  const setLoginUserData = () => {
+    const token = getToken();
+    if (!token) {
+      return navigate('/login');
+    }
+
+    const { userprofile } = jwtDecode(token);
+    if (!userprofile) return;
+
+    const loginUserProfile = {
+      memberIdx: userprofile.memberIdx,
+      id: userprofile.id,
+      profileImageUrl: userprofile.profileImageUrl,
+      name: userprofile.login,
+    };
+
+    userDispatch({ type: 'SET_USER', payload: loginUserProfile });
+  };
+
+  useEffect(() => {
+    setLoginUserData();
+  }, []);
 
   return (
     <>
