+++ frontend/src/containers/IssuePage/IssuePage.jsx
@@ -12,11 +12,6 @@ import { IssueList } from '@containers/index';
 import { isFilterApplied } from '@services/issue';
 import { useNavigate } from 'react-router-dom';
 
-import { getToken } from '@src/services/login';
-import { storeContext } from '@stores/index';
-import { useContext, useReducer } from 'react';
-import jwtDecode from 'jwt-decode';
-
 export const IssuePage = () => {
   const cx = classNames.bind(styles);
   const issuePageClassNames = `${cx('issue-page')}`;
@@ -35,26 +30,6 @@ export const IssuePage = () => {
   const [issueCount, setIssueCounts] = useState({ open: 0, closed: 0 });
 
   const navigate = useNavigate();
-  const [user, userDispatch] = useContext(storeContext).user;
-
-  const setLoginUserData = () => {
-    const token = getToken();
-    const { userprofile } = jwtDecode(token);
-    if (!userprofile) return;
-
-    const loginUserProfile = {
-      memberIdx: userprofile.memberIdx,
-      id: userprofile.id,
-      profileImageUrl: userprofile.profileImageUrl,
-      name: userprofile.login,
-    };
-
-    userDispatch({ type: 'SET_USER', payload: loginUserProfile });
-  };
-
-  useEffect(() => {
-    setLoginUserData();
-  }, []);
 
   useEffect(() => {
     const noneLabel = {
